move some components from PubPageContent to PubContent
